Bumped to version 0.1.4

+++ lib/core/file_merge.rb
@@ -22,7 +22,7 @@ module Expansions
 
     def run
       copy_name = "_copy_of#{File.basename(@output_file)}"
-      FileUtils.cp @output_file,copy_name
+      FileUtils.cp @output_file,copy_name if File.exist?(@output_file)
       FileUtils.rm_f @output_file
       File.open_for_write(@output_file) do |file|
         do_merge copy_name,file
@@ -32,7 +32,7 @@ module Expansions
 
     def do_merge(temp_file,target_file)
       merge_files(@before_files,target_file)
-      if @read_original_contents
+      if @read_original_contents && File.exist?(temp_file)
         File.open_for_read temp_file do|line|
           target_file << line
         end
